+++ packages/mantine/src/theme/plasmaCSSVariablesResolver.ts
@@ -1,14 +1,36 @@
-import {CSSVariablesResolver, ConvertCSSVariablesInput, keys} from '@mantine/core';
+import {alpha, ConvertCSSVariablesInput, CSSVariablesResolver} from '@mantine/core';
 
 export const plasmaCSSVariablesResolver: CSSVariablesResolver = (theme) => {
     const result: ConvertCSSVariablesInput = {
         variables: {},
         dark: {},
-        light: {},
+        light: {
+            // custom colors
+            '--coveo-color-input-border': theme.colors.gray[3],
+            '--coveo-color-title': theme.colors.gray[8],
+            '--coveo-color-text-disabled': theme.colors.gray[3],
+            '--coveo-color-bg-disabled': alpha(theme.colors.gray[4], 0.1),
+            '--coveo-color-text-readonly': 'var(--mantine-color-text)',
+            '--coveo-color-bg-readonly': theme.colors.gray[1],
+
+            // mantine overrides
+            '--mantine-color-default-border': theme.colors.gray[2],
+            '--mantine-color-error': theme.colors.red[5],
+            '--mantine-color-text': theme.colors.gray[6],
+            '--mantine-color-dimmed': theme.colors.gray[5],
+            '--mantine-color-gray-filled': theme.colors.gray[4],
+            '--mantine-color-gray-light': theme.colors.gray[0],
+            '--mantine-color-gray-light-hover': theme.colors.gray[1],
+            '--mantine-color-warning-filled': theme.colors.yellow[4],
+            '--mantine-color-placeholder': theme.colors.gray[4],
+            '--mantine-color-default-hover': theme.colors.gray[1],
+            /**
+             * ADUI-10725 https://coveord.atlassian.net/browse/ADUI-10725
+             * '--mantine-color-disabled'
+             * '--mantine-color-disabled-color'
+             * '--mantine-color-disabled-border'
+             */
+        },
     };
-    keys(theme.colors).forEach((color) => {
-        result.light[`--mantine-color-${color}-light`] = theme.colors[color][color === 'gray' ? 1 : 0];
-        result.light[`--mantine-color-${color}-light-hover`] = theme.colors[color][color === 'gray' ? 2 : 1];
-    });
     return result;
 };