+++ packages/website/src/building-blocs/Demo.tsx
@@ -3,15 +3,15 @@ import {
     Anchor,
     Box,
     Center,
+    CopyToClipboard,
     Flex,
     ScrollArea,
     SimpleGrid,
     Stack,
     Title,
     Tooltip,
-    useClipboard,
 } from '@coveord/plasma-mantine';
-import {CheckSize16Px, CopySize16Px, LinksSize16Px, PlaySize16Px} from '@coveord/plasma-react-icons';
+import {IconPlayerPlay, LinksSize16Px} from '@coveord/plasma-react-icons';
 import {CodeHighlight, CodeHighlightTabs} from '@mantine/code-highlight';
 import '@mantine/code-highlight/styles.css';
 import {Component, ReactNode} from 'react';
@@ -40,7 +40,6 @@ const Demo = ({
     maxHeight,
     additionalFiles,
 }: DemoProps) => {
-    const clipboard = useClipboard();
     const createSandbox = async () => {
         try {
             const res = await fetch(getCodeSandboxLink(snippet));
@@ -51,9 +50,9 @@ const Demo = ({
         }
     };
     return (
-        <div className={DemoClasses.root}>
+        <div>
             {title ? (
-                <Anchor href={`#${id}`} c="gray.9" className={DemoClasses.anchor}>
+                <Anchor href={`#${id}`} className={DemoClasses.anchor}>
                     <Title order={5} mb="xs" id={id} className={DemoClasses.title}>
                         {title}
                     </Title>
@@ -103,26 +102,11 @@ const Demo = ({
                         />
                     )}
                     <Stack className={DemoClasses.actions} gap="xs">
-                        <Tooltip label={clipboard.copied ? 'Copied' : 'Copy'} position="left">
-                            <ActionIcon
-                                variant="subtle"
-                                radius="sm"
-                                c={'var(--mantine-color-gray-6)'}
-                                onClick={() => clipboard.copy(snippet)}
-                            >
-                                {clipboard.copied ? <CheckSize16Px height={16} /> : <CopySize16Px height={16} />}
-                            </ActionIcon>
-                        </Tooltip>
-
+                        <CopyToClipboard color="gray" value={snippet} />
                         <Tooltip label="Open in CodeSanbox" position="left">
-                            <ActionIcon
-                                variant="subtle"
-                                radius="sm"
-                                c={'var(--mantine-color-gray-6)'}
-                                onClick={createSandbox}
-                            >
-                                <PlaySize16Px height={16} />
-                            </ActionIcon>
+                            <ActionIcon.Quaternary color="gray" onClick={createSandbox}>
+                                <IconPlayerPlay size={20} />
+                            </ActionIcon.Quaternary>
                         </Tooltip>
                     </Stack>
                 </div>