+++ packages/website/src/examples/layout/StickyFooter/StickyFooter.demo.tsx
@@ -2,8 +2,8 @@ import {Button, StickyFooter} from '@coveord/plasma-mantine';
 
 const Demo = () => (
     <StickyFooter borderTop>
-        <Button variant="outline">Cancel</Button>
-        <Button variant="filled">Save</Button>
+        <Button.Tertiary>Cancel</Button.Tertiary>
+        <Button.Primary>Save</Button.Primary>
     </StickyFooter>
 );
 export default Demo;